+++ src/Content/Backend/NV.Templates.Backend.Core/Framework/Continuation/ContinuationToken.cs
@@ -14,7 +14,7 @@ namespace NV.Templates.Backend.Core.Framework.Continuation
         /// Safe to transmit over HTTP/URI.
         /// </summary>
         /// <param name="value">The value to encode.</param>
-        /// <returns>The encoded continuation token</returns>
+        /// <returns>The encoded continuation token.</returns>
         public static string ToContinuationToken(object value)
         {
             return Convert.ToBase64String(Encoding.UTF8.GetBytes(JsonConvert.SerializeObject(value)));
@@ -23,9 +23,9 @@ namespace NV.Templates.Backend.Core.Framework.Continuation
         /// <summary>
         /// Recompose the <paramref name="continuationToken"/> into a <typeparamref name="T"/>.
         /// </summary>
-        /// <typeparam name="T">The continuation token decoded type</typeparam>
-        /// <param name="continuationToken">The encoded continuation token</param>
-        /// <returns>The decoded continuation token, or default if no continuation token,</returns>
+        /// <typeparam name="T">The continuation token decoded type.</typeparam>
+        /// <param name="continuationToken">The encoded continuation token.</param>
+        /// <returns>The decoded continuation token, or default if no continuation token,.</returns>
         public static T FromContinuationToken<T>(string continuationToken)
         {
             try