+++ CHANGELOG.md
@@ -26,6 +26,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Added LimitOffsetContinuationToken (ContinuationToken implementation using Limit & Offset pagination information)
+
 ### Changed
 
 ### Deprecated